Initial client: GetEvents only for now
