+++ cmd/stellar-rpc/internal/methods/simulate_transaction.go
@@ -17,6 +17,7 @@ import (
 	"github.com/stellar/stellar-rpc/cmd/stellar-rpc/internal/db"
 	"github.com/stellar/stellar-rpc/cmd/stellar-rpc/internal/preflight"
 	"github.com/stellar/stellar-rpc/cmd/stellar-rpc/internal/xdr2json"
+	"github.com/stellar/stellar-rpc/protocol"
 )
 
 type SimulateTransactionRequest struct {
@@ -90,7 +91,7 @@ func (l *LedgerEntryChangeType) UnmarshalJSON(data []byte) error {
 }
 
 func (l *LedgerEntryChange) FromXDRDiff(diff preflight.XDRDiff, format string) error {
-	if err := IsValidFormat(format); err != nil {
+	if err := protocol.IsValidFormat(format); err != nil {
 		return err
 	}
 
@@ -133,7 +134,7 @@ func (l *LedgerEntryChange) FromXDRDiff(diff preflight.XDRDiff, format string) e
 	}
 
 	switch format {
-	case FormatJSON:
+	case protocol.FormatJSON:
 		return l.jsonXdrDiff(diff, key)
 
 	default:
@@ -226,7 +227,7 @@ type PreflightGetter interface {
 // NewSimulateTransactionHandler returns a json rpc handler to run preflight simulations
 func NewSimulateTransactionHandler(logger *log.Entry, ledgerEntryReader db.LedgerEntryReader, ledgerReader db.LedgerReader, daemon interfaces.Daemon, getter PreflightGetter) jrpc2.Handler {
 	return NewHandler(func(ctx context.Context, request SimulateTransactionRequest) SimulateTransactionResponse {
-		if err := IsValidFormat(request.Format); err != nil {
+		if err := protocol.IsValidFormat(request.Format); err != nil {
 			return SimulateTransactionResponse{Error: err.Error()}
 		}
 
@@ -315,7 +316,7 @@ func NewSimulateTransactionHandler(logger *log.Entry, ledgerEntryReader db.Ledge
 		var results []SimulateHostFunctionResult
 		if len(result.Result) != 0 {
 			switch request.Format {
-			case FormatJSON:
+			case protocol.FormatJSON:
 				rvJs, err := xdr2json.ConvertBytes(xdr.ScVal{}, result.Result)
 				if err != nil {
 					return SimulateTransactionResponse{
@@ -350,7 +351,7 @@ func NewSimulateTransactionHandler(logger *log.Entry, ledgerEntryReader db.Ledge
 		var restorePreamble *RestorePreamble = nil
 		if len(result.PreRestoreTransactionData) != 0 {
 			switch request.Format {
-			case FormatJSON:
+			case protocol.FormatJSON:
 				txDataJs, err := xdr2json.ConvertBytes(
 					xdr.SorobanTransactionData{},
 					result.PreRestoreTransactionData)
@@ -394,7 +395,7 @@ func NewSimulateTransactionHandler(logger *log.Entry, ledgerEntryReader db.Ledge
 		}
 
 		switch request.Format {
-		case FormatJSON:
+		case protocol.FormatJSON:
 			simResp.TransactionDataJSON, err = xdr2json.ConvertBytes(
 				xdr.SorobanTransactionData{},
 				result.TransactionData)
